+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2]
+### Changed
+* updated plugin helpers #27280
+
+### Fixed
+* eslint configuration #27280
+
 ## [0.2.1]
 ### Changed
 * prevent navigation on title text click